+++ static/app/views/dashboards/widgets/timeSeriesWidget/splitSeriesIntoCompleteAndIncomplete.tsx
@@ -2,14 +2,17 @@ import partition from 'lodash/partition';
 
 import type {TimeseriesData} from '../common/types';
 
+import {markDelayedData} from './markDelayedData';
+
 export function splitSeriesIntoCompleteAndIncomplete(
   serie: TimeseriesData,
   delay: number
 ): Array<TimeseriesData | undefined> {
   const markedTimeserie = markDelayedData(serie, delay);
 
-  const [completeData, incompleteData] = partition(markedTimeserie.data, datum =>
-    Boolean(datum.delayed)
+  const [completeData, incompleteData] = partition(
+    markedTimeserie.data,
+    datum => !datum.delayed
   );
 
   // If there is both complete and incomplete data, prepend the incomplete data
@@ -37,40 +40,6 @@ export function splitSeriesIntoCompleteAndIncomplete(
         }
       : undefined,
   ];
-}
-
-/**
- * Given a timeseries and a delay in seconds, goes through the timeseries data, and marks each point as either delayed (data bucket ended before the delay threshold) or not
- */
-function markDelayedData(timeserie: TimeseriesData, delay: number) {
-  if (delay === 0) {
-    return timeserie;
-  }
-
-  const penultimateDatum = timeserie.data.at(-2);
-  const finalDatum = timeserie.data.at(-1);
-
-  let bucketSize: number = 0;
-  if (penultimateDatum && finalDatum) {
-    bucketSize =
-      new Date(finalDatum.timestamp).getTime() -
-      new Date(penultimateDatum.timestamp).getTime();
-  }
-
-  const ingestionDelayTimestamp = Date.now() - delay * 1000;
-
-  return {
-    ...timeserie,
-    data: timeserie.data.map(datum => {
-      const bucketEndTimestamp = new Date(datum.timestamp).getTime() + bucketSize;
-      const delayed = bucketEndTimestamp < ingestionDelayTimestamp;
-
-      return {
-        ...datum,
-        delayed,
-      };
-    }),
-  };
 }
 
 function discardDelayProperty(
